move UiSelector and BySelector extensions to ContractsExtensions.kt
